// server/pay/types.ts
export type UnifiedPayStatus = 'WP' | 'OD' | 'CD' | 'CL' | 'FL' | 'RF'